Update zombie message to be more descriptive (#26141)
Because the zombie detection method comes from the scheduler job, in a distributed environment it can be cumbersome to find the worker that actually had the problem that led to the zombie. In editing this, I also noticed that the zombie message had some redundant and unhelpful elements (e.g. printing the memory address of the SimpleTaskInstance) so I've cleaned it up.

closes: #26067