Speed up provider validation pre-commit (#28541)
The provider validation pre-commit is now a lot slower after #28516
but it turned out that it has been also doing a little too much.

It worked in the way (against the original design) that when any
provider.yaml changed, validation was actually performed for all
the providers, not only for those changed.

This change improves the regular speed of running this validation
when one or few providers are changed in the commit.

It's not only limiting the validation to those provider.yaml files
that changes but also performing subset of validations on those changed
files if not all providers are changed or --all-files are not used
(because some of the validations involving documentation and
cross-provider dependencies require all provider.yaml files to be loaded
and validated).

The validation is still done in one "docker run" command though.