Apply "unify bucket and key" before "provide bucket" (#28710)
If we unify first, then we (desirably) get the bucket from the key.  If we provide bucket first, something undesirable happens.  The bucket from the connection is used, which means we don't respect the full key provided. Further, the full key is not made relative, which would cause the actual request to fail. For this reason we want to put unify first.  Because in the case where the outcome is different, the previously the request would fail, this should not pose any backcompat concerns.

Co-authored-by: Felix Uellendall <feluelle@users.noreply.github.com>