Use entire tenant domain name in dbt Cloud connection (#28890)
Closes: #27146

Currently users can connect to both single-tenant and multi-tenant instances of dbt Cloud. However, for the enterprise tier, there is a new region offered as well as private instance support. The existing `tenant` value does not cover the ability to modify the entire domain name which could be required for users on the enterprise plan.

This enhancement allows users to specify the tenant domain they wish to connect to while keeping the "cloud.getdbt.com" domain as the default. Additionally, backward compat is implemented for those who are only specifying the third-level domain for single-tenant instances and wish to upgrade the provider version.