Upgrade redis (#31279)
Upgrade redis to 4.5.5

Upgrading redis package >= 4 doesn't require code changes. Changes in Redis 4 only brings into impact when we are directly importing redis.client to execute the commands, which we are not doing. The issue with unquoting URLS with urllib.parse.unquote is also not applicable to airflow, because we are not using the from_url function