Fix for TaskGroup toggles for duplicated labels (#34072)
* Fix for TaskGroup toggles for duplicated labels

We have been using `_.label` so far to remember which task groups are opened/closed. Unfortunately, this leads to issues when the same label is used twice, which can happen as described in the linked issue.
With this change we switch to use `_.id` instead, which must be unique (otherwise parsing the DAG would lead to `DuplicateTaskIdFound`).

Fixes #34066

* Fix toggle for Gantt as well